[Savestates] Refactored SortedArrayList
